requirements.txt
streamlit==1.33.0
requests==2.32.3
google-generativeai==0.8.3